Artifact Store RDBS: Renaming modules to match current concept terms.
